Aligning with the grid now no longer interrupts the walking animation if an appropriate key is still held down.

+++ source/seed.c
@@ -19,11 +19,11 @@ int32_t	seed_get(t_game *game)
 	int32_t	a;
 	int32_t	b;
 	int32_t	c;
-	int32_t	d;
+	int64_t	d;
 
 	a = game->map->dims.h * game->map->dims.w;
 	b = game->map->player->position.x * game->map->player->position.y;
 	c = (int32_t) game->score_max;
-	d = (int32_t) game->mlx;
+	d = (int64_t) game->mlx;
 	return (d / (a - b + c) * (d % (a + c)));
 }
